Async#complete() throws an IllegalStateException when called more than once
