+++ src/main/java/io/vertx/ext/unit/Async.java
@@ -11,11 +11,9 @@ import io.vertx.codegen.annotations.VertxGen;
 public interface Async {
 
   /**
-   * Signals the asynchronous operation is done, this method should be called only once, calling it several
-   * times is tolerated.
-   *
-   * @return true when called the first time, false otherwise.
+   * Signals the asynchronous operation is done, this method should be called only once, if the method is called
+   * another time it will throw an {@code IllegalStateException} to signal the error.
    */
-  boolean complete();
+  void complete();
 
 }
