+++ src/main/java/io/vertx/ext/unit/impl/InvokeTask.java
@@ -1,13 +1,9 @@
 package io.vertx.ext.unit.impl;
 
 import io.vertx.core.Handler;
-import io.vertx.core.Vertx;
-import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.Test;
 import io.vertx.ext.unit.report.TestResult;
-import org.junit.Assert;
 
-import java.util.LinkedList;
 import java.util.concurrent.TimeoutException;
 
 /**
@@ -15,9 +11,9 @@ import java.util.concurrent.TimeoutException;
 */
 class InvokeTask implements Task<Result> {
 
-  private final Task<Result> next;
-  private final Handler<Test> test;
-  private final long timeout;
+  final Task<Result> next;
+  final Handler<Test> test;
+  final long timeout;
 
   public InvokeTask(Handler<Test> test, Task<Result> next) {
     this.test = test;
@@ -33,196 +29,8 @@ class InvokeTask implements Task<Result> {
 
   @Override
   public void execute(Result failure, Context context) {
-
-    class TestImpl implements Test {
-
-      private boolean completed;
-      private Throwable failed;
-      private long time;
-      private final LinkedList<Async> asyncs = new LinkedList<>();
-
-      public TestImpl(long time, Throwable failed) {
-        this.failed = failed;
-        this.time = time - System.currentTimeMillis();
-      }
-
-      private void tryEnd() {
-        boolean run = false;
-        synchronized (this) {
-          if (asyncs.isEmpty() && !completed) {
-            completed = true;
-            run = true;
-            time += System.currentTimeMillis();
-          }
-        }
-        if (run) {
-          context.run(next, new Result(time, failed));
-        }
-      }
-
-      private void failed(Throwable t) {
-        synchronized (this) {
-          if (!completed) {
-            failed = t;
-          }
-        }
-        while (true) {
-          Async async;
-          synchronized (this) {
-            async = asyncs.peekFirst();
-          }
-          if (async == null) {
-            break;
-          } else {
-            async.complete();
-          }
-        }
-      }
-
-      @Override
-      public Vertx vertx() {
-        return context.vertx();
-      }
-
-      @Override
-      public Async async() {
-        synchronized (this) {
-          if (!completed) {
-            Async async = new Async() {
-              @Override
-              public boolean complete() {
-                boolean completed;
-                synchronized (TestImpl.this) {
-                  completed = asyncs.remove(this);
-                }
-                if (completed) {
-                  tryEnd();
-                  return true;
-                } else {
-                  return false;
-                }
-              }
-            };
-            asyncs.add(async);
-            return async;
-          } else {
-            throw new IllegalStateException("Test already completed");
-          }
-        }
-      }
-
-      @Override
-      public Test assertTrue(boolean condition, String message) {
-        try {
-          Assert.assertTrue(message, condition);
-          return this;
-        } catch (AssertionError err) {
-          failed(err);
-          throw err;
-        }
-      }
-
-      public Test assertTrue(boolean condition) {
-        try {
-          Assert.assertTrue(condition);
-          return this;
-        } catch (AssertionError err) {
-          failed(err);
-          throw err;
-        }
-      }
-
-      public void fail(String message) {
-        try {
-          Assert.fail(message);
-        } catch (AssertionError err) {
-          failed(err);
-          throw err;
-        }
-      }
-
-      @Override
-      public Test assertFalse(boolean condition) {
-        try {
-          Assert.assertFalse(condition);
-          return this;
-        } catch (AssertionError err) {
-          failed(err);
-          throw err;
-        }
-      }
-
-      @Override
-      public Test assertFalse(boolean condition, String message) {
-        try {
-          Assert.assertFalse(message, condition);
-          return this;
-        } catch (AssertionError err) {
-          failed(err);
-          throw err;
-        }
-      }
-
-      @Override
-      public void fail() {
-        try {
-          Assert.fail();
-        } catch (AssertionError err) {
-          failed(err);
-          throw err;
-        }
-      }
-
-      @Override
-      public Test assertEquals(Object expected, Object actual) {
-        try {
-          Assert.assertEquals(expected, actual);
-          return this;
-        } catch (AssertionError err) {
-          failed(err);
-          throw err;
-        }
-      }
-
-      @Override
-      public Test assertEquals(Object expected, Object actual, String message) {
-        try {
-          Assert.assertEquals(message, expected, actual);
-          return this;
-        } catch (AssertionError err) {
-          failed(err);
-          throw err;
-        }
-      }
-
-      @Override
-      public Test assertNotEquals(Object first, Object second, String message) {
-        try {
-          Assert.assertNotEquals(message, first, second);
-          return this;
-        } catch (AssertionError err) {
-          failed(err);
-          throw err;
-        }
-      }
-
-      @Override
-      public Test assertNotEquals(Object first, Object second) {
-        try {
-          Assert.assertNotEquals(first, second);
-          return this;
-        } catch (AssertionError err) {
-          failed(err);
-          throw err;
-        }
-      }
-    };
-
-    //
-    TestImpl test = new TestImpl(failure != null ? failure.time : 0, failure != null ? failure.failure : null);
-    Async async = test.async();
-
-    //
+    TestImpl test = new TestImpl(this, context, failure != null ? failure.time : 0, failure != null ? failure.failure : null);
+    TestImpl.AsyncImpl async = test.async();
     if (timeout > 0) {
       Runnable cancel = () -> {
         try {
@@ -238,13 +46,12 @@ class InvokeTask implements Task<Result> {
         new Thread(cancel).start();
       }
     }
-
     try {
       InvokeTask.this.test.handle(test);
     } catch (Throwable t) {
       test.failed(t);
     } finally {
-      async.complete();
+      async.internalComplete();
     }
   }
